update in postcontroller and commentcontroller

+++ Developers_Oasis/app/Http/Controllers/CommentController.php
@@ -7,6 +7,8 @@ use App\Models\Comment;
 use App\Http\Requests\StoreCommentRequest;
 use App\Http\Requests\UpdateCommentRequest;
 use App\Http\Resources\CommentResource;
+use Exception;
+use Illuminate\Http\Response;
 
 
 class CommentController extends Controller
@@ -31,13 +33,32 @@ class CommentController extends Controller
 
     public function update(UpdateCommentRequest $request, Comment $comment)
     {
-          $comment->update($request->all());
-          return new CommentResource($comment);
+        if($comment)
+        try {
+            $comment->update($request->all());
+            return new CommentResource($comment);
+
+        }catch(Exception $e ){
+            return $e;
+        }else{
+            return response('', 404);
+        }
     }
 
     public function destroy(Comment $comment)
     {
-        $comment->delete();
-        return response()->json(['message' => 'Comment deleted successfully']);
+        if($comment){
+            try{
+                $comment->delete();
+                return new Response('', 204);
+            }catch(Exception $e){
+                return $e;
+
+            }
+        }else{
+            return response()->json('', 404);
+        }
+
     }
-}
+    }
+
